requirements.txt
Flask==2.3.2
Flask_SocketIO==5.3.6
opencv_contrib_python==4.8.0.74
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.0.0
Pillow==10.2.0
